refactor(cli): one kind-to-input reader and one unreadable guard for Copilot and OpenClaw
`READABLE`, `unreadableInput` and `preGuards` were line-for-line identical in
`adapters/copilot.ts` and `adapters/openclaw.ts` modulo renames, and the
kind-to-record readers in `copilot-input.ts` and `openclaw-input.ts` mirrored
each other down to the path and URL candidate helpers. A copy of a security
check is a fix that lands on one agent only — and `withCandidates`, which
stops a payload's own `urls`/`file_paths` from choosing what gets judged, is
exactly such a check.

`adapters/kind-input.ts` now owns the shared `ToolKind`, the path/URL
candidate readers, `kindToolInput`, `preCandidatesFor` and `unreadableGuard`.
Each adapter keeps only what genuinely differs: which name maps to which
kind, which keys its file tools drop, and how its deny is worded.
Behaviour-preserving — no test file changed, and all 1341 stay green.
